test(sasjs-cli): moved specs into dedicated folder
