+++ src/sasjsCli/specs/testFiles/binaries/1x1.png
