+++ src/sasjsCli/specs/testFiles/binaries/img.png
