+++ src/sasjsCli/specs/testFiles/exampleWithBinaries-2.sas
