+++ src/sasjsCli/specs/testFiles/exampleWithBinaries.sas
